Skip updating `conda` (on AppVeyor)
For some strange reason `conda` 4.2.13 seems willing to do **anything**
to upgrade itself. This includes downgrading from Python 3 to Python 2
in order to get the latest `conda`. The result is we actually can't
build and release `conda` for Python 3.5 and 3.6 currently.

To fix this, we hack the AppVeyor script to disable the `conda` update
step. Also configure `conda` to disable its auto-updating feature. So it
won't try to update itself while doing something else. Hopefully this is
sufficient to workaround the problem. A re-rendering will clear out this
hack once we have completed the release.