Construct a more natural environment to test conda
Attempts to sort of emulate a `root`-like environment for `conda` to be
in for testing. It is not truly the `root` environment though as it is
still managed by `conda` from the actual `root` environment and this
does show up in a couple of spots.

Some of the things done include specifying a separate package cache,
environment directory, and `.condarc` within the test environment.
Further the actual `root` environment is scrubbed from the path. This is
all done with a careful eye towards protecting a user that tries running
these tests on their own machine so as to avoid messing up their `conda`
install. Still users should approach this cautiously as this is very
experimental at this point. In short this tries as much as possible to
mimic a normal `root`-installed `conda` in as clean a way as possible.

Adding this framework should make it easier to test things like
activating and deactivating environments, create and removing
environments, installing and uninstalling packages, etc. All in all this
should allow us to have more confidence in `conda` packages before
releasing them to `conda-forge`.